requirements.txt
python-dotenv
undetected-chromedriver
setuptools
